requirements.txt
adafruit-circuitpython-neopixel
spotipy
requests
flask
rpi_ws281x